requirements.txt
# Super AI Project Dependencies

# Core requirements
flask==2.2.3
werkzeug==2.2.3
requests==2.28.2
beautifulsoup4==4.11.2
pydantic==1.10.7

# Security dependencies
flask-jwt-extended==4.4.4
flask-cors==3.0.10
python-magic==0.4.27
cryptography==39.0.2

# Performance optimization dependencies
flask-caching==2.0.2
pillow==9.4.0
celery==5.2.7
redis==4.5.1
gunicorn==20.1.0
whitenoise==6.4.0

# Additional dependencies
python-dotenv==1.0.0

# Deep learning
torch>=2.0.0
torchvision>=0.15.0
# Uncomment if GPU support is needed
# torchaudio>=0.9.0

# Async support
aiohttp>=3.8.0
asyncio>=3.4.3

# Web interface
Jinja2>=3.0.0
itsdangerous>=2.0.0
click>=8.0.0

# Utilities
tqdm>=4.62.0
matplotlib>=3.4.0
PyYAML>=6.0

# Optional: For production deployment
# gunicorn>=20.1.0
# uvicorn>=0.15.0

# Web and API
flask>=2.0.0
requests>=2.28.0
aiohttp>=3.8.0

# Data processing
pandas>=1.4.0
numpy>=1.22.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Machine learning
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.1.0

# Utilities
aiofiles>=0.8.0

# Development tools
pytest>=7.0.0
pytest-asyncio>=0.18.0
black>=22.1.0

# Deployment options (uncomment as needed)
# gunicorn>=20.1.0
# uvicorn>=0.15.0
